Implement card-badges and card-detail-badges
For the badges we look at the title of the card to determine what kind
of badge to color. If the card has 'green', 'yellow', or 'red' in the
name we color the badge accordingly. Additionally, if the card has
dynamic in the name, we will use a dynamic badge instead of a static
one.